Refactored data integration tests
