+++ test/data/integration/targets.js
@@ -1,0 +1,2 @@
+// we can use this to read configuration from file or command line to specify which providers to test
+module.exports = ['mssql'];
