--- test/data/mssql/integration.schema.tests.js
@@ -1,62 +1,0 @@
-// ----------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// ----------------------------------------------------------------------------
-var config = require('../../appFactory').configuration().data,
-    schema = require('../../../src/data/mssql/schema')(config),
-    statements = require('../../../src/data/mssql/statements'),
-    queries = require('../../../src/query'),
-    execute = require('../../../src/data/mssql/execute'),
-    expect = require('chai')
-        .use(require('chai-subset'))
-        .use(require('chai-as-promised'))
-        .expect,
-    table = { name: 'schemaTest', seed: [{ id: '1' }, { id: '2' }] };
-
-describe('azure-mobile-apps.data.sql.integration.schema', function () {
-    // schema was refactored out of dynamicSchema. Most of the functionality is tested by the dynamicSchema tests.
-    afterEach(function (done) {
-        execute(config, { sql: 'drop table dbo.schemaTest' }).then(done, done);
-    });
-
-    describe('initialize', function () {
-        it("initialize creates table and seeds", function () {
-            return schema.initialize(table)
-                .then(function () {
-                    return execute(config, statements.read(queries.create('schemaTest')));
-                })
-                .then(function (results) {
-                    expect(results.length).to.equal(2);
-                })
-        });
-
-        it("initialize handles existing tables", function () {
-            return schema.initialize(table)
-                .then(function () {
-                    return schema.initialize(table);
-                })
-                .then(function () {
-                    return execute(config, statements.read(queries.create('schemaTest')));
-                })
-                .then(function (results) {
-                    expect(results.length).to.equal(2);
-                })
-        });
-    });
-
-    describe('get', function () {
-        it("returns databse schema for table", function () {
-            return schema.initialize(table)
-                .then(function () {
-                    return schema.get(table);
-                })
-                .then(function (tableSchema) {
-                    expect(tableSchema).to.containSubset({
-                        name: table.name,
-                        properties: [
-                            { name: 'id', type: 'string' }
-                        ]
-                    });
-                });
-        });
-    });
-});
